refactored parity tests to support all providers
changed scope from session to module

moved fixtures to conftest

fixed linting issues

refactored parity tests

fixed circle ci

removed unused import
