+++ .circleci/config.yml
@@ -40,7 +40,7 @@ parity_steps: &parity_steps
         command: |
           pip install --user requests eth_utils tqdm eth-hash[pycryptodome]
           echo $PARITY_VERSION
-          python tests/integration/install_parity.py $PARITY_VERSION
+          python tests/integration/parity/install_parity.py $PARITY_VERSION
     - run:
         name: run tox
         command: ~/.local/bin/tox -r
@@ -132,12 +132,20 @@ jobs:
       TOXENV: py35-integration-goethereum
       GETH_VERSION: v1.8.1
 
-  py35-integration-parity:
+  py35-integration-parity-ipc:
     <<: *parity_steps
     docker:
       - image: circleci/python:3.5
     environment:
-      TOXENV: py35-integration-parity
+      TOXENV: py35-integration-parity-ipc
+      PARITY_VERSION: v1.8.8
+
+  py35-integration-parity-http:
+    <<: *parity_steps
+    docker:
+      - image: circleci/python:3.5
+    environment:
+      TOXENV: py35-integration-parity-http
       PARITY_VERSION: v1.8.8
 
   py35-integration-ethtester-pyethereum:
@@ -202,12 +210,20 @@ jobs:
       TOXENV: py36-integration-goethereum
       GETH_VERSION: v1.8.1
 
-  py36-integration-parity:
+  py36-integration-parity-ipc:
     <<: *parity_steps
     docker:
       - image: circleci/python:3.6
     environment:
-      TOXENV: py36-integration-parity
+      TOXENV: py36-integration-parity-ipc
+      PARITY_VERSION: v1.8.8
+
+  py36-integration-parity-http:
+    <<: *parity_steps
+    docker:
+      - image: circleci/python:3.6
+    environment:
+      TOXENV: py36-integration-parity-http
       PARITY_VERSION: v1.8.8
 
   py36-integration-ethtester-pyethereum:
@@ -236,13 +252,15 @@ workflows:
       - py35-ens
       - py35-integration-goethereum-1.7.2
       - py35-integration-goethereum-1.8.1
-      - py35-integration-parity
+      - py35-integration-parity-ipc
+      - py35-integration-parity-http
       - py35-integration-ethtester-pyethereum
       - py35-integration-ethtester-pyevm
       - py36-core
       - py36-ens
       - py36-integration-goethereum-1.7.2
       - py36-integration-goethereum-1.8.1
-      - py36-integration-parity
+      - py36-integration-parity-ipc
+      - py36-integration-parity-http
       - py36-integration-ethtester-pyethereum
       - py36-integration-ethtester-pyevm
