+++ .travis.yml
@@ -15,9 +15,12 @@ matrix:
       env: TOX_POSARGS="-e py35-integration-goethereum" GETH_VERSION=v1.7.2
     - python: "3.5"
       env: TOX_POSARGS="-e py35-integration-goethereum" GETH_VERSION=v1.8.1
-    # parity
+    # parity-ipc
     - python: "3.5"
-      env: TOX_POSARGS="-e py35-integration-parity" PARITY_VERSION=v1.8.8
+      env: TOX_POSARGS="-e py35-integration-parity-ipc" PARITY_VERSION=v1.8.8
+    # parity-http
+    - python: "3.5"
+      env: TOX_POSARGS="-e py35-integration-parity-http" PARITY_VERSION=v1.8.8
     # ENS
     - python: "3.5"
       env: TOX_POSARGS="-e py35-ens"
@@ -38,9 +41,12 @@ matrix:
       env: TOX_POSARGS="-e py36-integration-goethereum" GETH_VERSION=v1.7.2
     - python: "3.6"
       env: TOX_POSARGS="-e py36-integration-goethereum" GETH_VERSION=v1.8.1
-    # parity
+    # parity-ipc
     - python: "3.6"
-      env: TOX_POSARGS="-e py36-integration-parity" PARITY_VERSION=v1.8.8
+      env: TOX_POSARGS="-e py36-integration-parity-ipc" PARITY_VERSION=v1.8.8
+    # parity-http
+    - python: "3.6"
+      env: TOX_POSARGS="-e py36-integration-parity-http" PARITY_VERSION=v1.8.8
     # ENS
     - python: "3.6"
       env: TOX_POSARGS="-e py36-ens"
@@ -83,7 +89,7 @@ before_script:
   - if [ -n "$GETH_VERSION" ] && [ ! -e "$GETH_BINARY" ]; then python -m geth.install $GETH_VERSION; fi
   - if [ -n "$GETH_VERSION" ]; then $GETH_BINARY version; fi
   - if [ -n "$GETH_VERSION" ]; then $GETH_BINARY makedag 0 $HOME/.ethash; fi
-  - if [ -n "$PARITY_VERSION" ]; then python tests/integration/install_parity.py $PARITY_VERSION; fi
+  - if [ -n "$PARITY_VERSION" ]; then python tests/integration/parity/install_parity.py $PARITY_VERSION; fi
   - pip freeze
 script:
   - tox $TOX_POSARGS
