+++ tests/integration/parity/common.py
@@ -1,0 +1,110 @@
+import pytest
+
+from web3.utils.module_testing import (
+    EthModuleTest,
+    PersonalModuleTest,
+    Web3ModuleTest,
+)
+
+
+class ParityWeb3ModuleTest(Web3ModuleTest):
+    def _check_web3_clientVersion(self, client_version):
+        assert client_version.startswith('Parity/')
+
+
+class ParityEthModuleTest(EthModuleTest):
+    def test_eth_uninstallFilter(self, web3):
+        pytest.xfail('eth_uninstallFilter calls to parity always return true')
+        super().test_eth_uninstallFilter(web3)
+
+    def test_eth_newBlockFilter(self, web3):
+        pytest.xfail('Parity returns latest block on first polling for new blocks')
+        super().test_eth_newBlockFilter(web3)
+
+    def test_eth_replaceTransaction(self, web3, unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_replaceTransaction(web3, unlocked_account)
+
+    def test_eth_replaceTransaction_already_mined(self, web3, unlocked_account):
+        pytest.xfail('Parity is not setup to auto mine')
+        super().test_eth_replaceTransaction_already_mined(web3, unlocked_account)
+
+    def test_eth_replaceTransaction_incorrect_nonce(self, web3, unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_replaceTransaction_incorrect_nonce(web3, unlocked_account)
+
+    def test_eth_replaceTransaction_gas_price_too_low(self, web3, unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_replaceTransaction_gas_price_too_low(web3, unlocked_account)
+
+    def test_eth_replaceTransaction_gas_price_defaulting_minimum(self, web3, unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_replaceTransaction_gas_price_defaulting_minimum(web3, unlocked_account)
+
+    def test_eth_replaceTransaction_gas_price_defaulting_strategy_higher(self,
+                                                                         web3,
+                                                                         unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_replaceTransaction_gas_price_defaulting_strategy_higher(
+            web3, unlocked_account
+        )
+
+    def test_eth_replaceTransaction_gas_price_defaulting_strategy_lower(self,
+                                                                        web3,
+                                                                        unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_replaceTransaction_gas_price_defaulting_strategy_lower(
+            web3, unlocked_account
+        )
+
+    def test_eth_modifyTransaction(self, web3, unlocked_account):
+        pytest.xfail('Needs ability to efficiently control mining')
+        super().test_eth_modifyTransaction(web3, unlocked_account)
+
+
+class ParityPersonalModuleTest(PersonalModuleTest):
+    def test_personal_importRawKey(self, web3):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_importRawKey(web3)
+
+    def test_personal_listAccounts(self, web3):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_listAccounts(web3)
+
+    def test_personal_lockAccount(self, web3, unlocked_account):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_lockAccount(web3, unlocked_account)
+
+    def test_personal_unlockAccount_success(self, web3):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_unlockAccount_success(web3)
+
+    def test_personal_unlockAccount_failure(self, web3, unlockable_account):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_unlockAccount_failure(web3, unlockable_account)
+
+    def test_personal_newAccount(self, web3):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_newAccount(web3)
+
+    def test_personal_sendTransaction(
+            self,
+            web3,
+            unlockable_account,
+            unlockable_account_pw):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_sendTransaction(
+            web3,
+            unlockable_account,
+            unlockable_account_pw)
+
+    def test_personal_sign_and_ecrecover(
+            self,
+            web3,
+            unlockable_account,
+            unlockable_account_pw):
+        pytest.xfail('this non-standard json-rpc method is not implemented on parity')
+        super().test_personal_sign_and_ecrecover(
+            web3,
+            unlockable_account,
+            unlockable_account_pw)
