+++ tests/integration/parity/install_parity.py
