+++ tests/integration/parity/utils.py
@@ -1,0 +1,90 @@
+import signal
+import socket
+import subprocess
+import time
+
+from eth_utils import (
+    to_text,
+)
+import requests
+
+
+def wait_for_socket(ipc_path, timeout=60):
+    start = time.time()
+    while time.time() < start + timeout:
+        try:
+            sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
+            sock.connect(ipc_path)
+            sock.settimeout(timeout)
+        except (FileNotFoundError, socket.error):
+            time.sleep(0.01)
+        else:
+            break
+
+
+def wait_for_http(endpoint_uri, timeout=60):
+    start = time.time()
+    while time.time() < start + timeout:
+        try:
+            requests.get(endpoint_uri)
+        except requests.ConnectionError:
+            time.sleep(0.01)
+        else:
+            break
+
+
+def get_process(command_list, terminates=False):
+
+    proc = subprocess.Popen(
+        command_list,
+        stdin=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        bufsize=1,
+    )
+    if terminates:
+        wait_for_popen(proc, 30)
+    try:
+        yield proc
+    finally:
+        kill_proc_gracefully(proc)
+        output, errors = proc.communicate()
+        print(
+            "Parity Process Exited:\n"
+            "stdout:{0}\n\n"
+            "stderr:{1}\n\n".format(
+                to_text(output),
+                to_text(errors),
+            )
+        )
+
+
+def wait_for_popen(proc, timeout):
+    start = time.time()
+    while time.time() < start + timeout:
+        if proc.poll() is None:
+            time.sleep(0.01)
+        else:
+            break
+
+
+def kill_proc_gracefully(proc):
+    if proc.poll() is None:
+        proc.send_signal(signal.SIGINT)
+        wait_for_popen(proc, 13)
+
+    if proc.poll() is None:
+        proc.terminate()
+        wait_for_popen(proc, 5)
+
+    if proc.poll() is None:
+        proc.kill()
+        wait_for_popen(proc, 2)
+
+
+def get_open_port():
+    sock = socket.socket()
+    sock.bind(('127.0.0.1', 0))
+    port = sock.getsockname()[1]
+    sock.close()
+    return str(port)
