+++ tox.ini
@@ -30,7 +30,8 @@ commands=
     ens: py.test {posargs:tests/ens}
     integration-goethereum: py.test {posargs:tests/integration/test_goethereum.py}
     integration-ethtester: py.test {posargs:tests/integration/test_ethereum_tester.py}
-    integration-parity: py.test -s {posargs:tests/integration/test_parity.py}
+    integration-parity-ipc: py.test -s {posargs:tests/integration/parity/test_parity_ipc.py}
+    integration-parity-http: py.test -s {posargs:tests/integration/parity/test_parity_http.py}
     doctest: make -C {toxinidir}/docs doctest
 deps =
     -r{toxinidir}/requirements-dev.txt
